Staging

+++ app/Http/Controllers/AdminVideosController.php
@@ -1016,27 +1016,41 @@ if(!empty($artistsdata)){
                $player_image = $video->image;
            }
 
-        $path = public_path().'/uploads/videos/';
+           $video->trailer_type = $data['trailer_type'];
 
-           if(!empty($trailer)) {   
-            //code for remove old file
-            if($trailer != ''  && $trailer != null){
-                 $file_old = $path.$trailer;
-                if (file_exists($file_old)){
-                 unlink($file_old);
-                }
-            }
-            //upload new file
-            $randval = Str::random(16);
-            $file = $trailer;
-            $trailer_vid  = $randval.'.'.$request->file('trailer')->extension();
-            $file->move($path, $trailer_vid);
-            $data['trailer']  = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
-               $video->trailer = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
-            } else {
-                $data['trailer'] = $video->trailer;
-            }  
-        
+           // dd($data);
+   
+           if($data['trailer_type'] == 'video_mp4'){
+               if(!empty($trailer)) {   
+                   //code for remove old file
+                   if($trailer != ''  && $trailer != null){
+                        $file_old = $path.$trailer;
+                       if (file_exists($file_old)){
+                        unlink($file_old);
+                       }
+                   }
+                   //upload new file
+                   $randval = Str::random(16);
+                   $file = $trailer;
+                   $trailer_vid  = $randval.'.'.$request->file('trailer')->extension();
+                   $file->move($path, $trailer_vid);
+                   $data['trailer']  = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
+                      $video->trailer = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
+                   } else {
+                       $data['trailer'] = $video->trailer;
+                   }  
+   
+                   }elseif($data['trailer_type'] == 'm3u8_url'){
+                       $video->trailer = $data['m3u8_trailer'];
+                   }
+                   elseif($data['trailer_type'] == 'mp4_url'){
+                       $video->trailer = $data['mp4_trailer'];
+                   }
+                   elseif($data['trailer_type'] == 'embed_url'){
+                       $video->trailer = $data['embed_trailer'];
+                   }
+                   
+           
            $update_mp4 = $request->get('video');
            if(empty($data['active'])){
             $active = 0;
@@ -1928,24 +1942,37 @@ if(!empty($artistsdata)){
              }
             
             
-            if($trailer != '') {   
-                  //code for remove old file
-                  if($trailer != ''  && $trailer != null){
-                       $file_old = $path.$trailer;
-                      if (file_exists($file_old)){
-                       unlink($file_old);
-                      }
-                  }
-                  //upload new file
-                  $randval = Str::random(16);
-                  $file = $trailer;
-                  $trailer_vid  = $randval.'.'.$request->file('trailer')->extension();
-                  $file->move($path, $trailer_vid);
-                  $data['trailer']  = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
-    
-             } else {
-                 $data['trailer'] = $video->trailer;
-             }  
+             $video->trailer_type = $data['trailer_type'];
+
+             if($data['trailer_type'] == 'video_mp4'){
+                 if($trailer != '') {   
+                     //code for remove old file
+                     if($trailer != ''  && $trailer != null){
+                          $file_old = $path.$trailer;
+                         if (file_exists($file_old)){
+                          unlink($file_old);
+                         }
+                     }
+                     //upload new file
+                     $randval = Str::random(16);
+                     $file = $trailer;
+                     $trailer_vid  = $randval.'.'.$request->file('trailer')->extension();
+                     $file->move($path, $trailer_vid);
+                     $data['trailer']  = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
+       
+                } else {
+                    $data['trailer'] = $video->trailer;
+                }  
+ 
+             }elseif($data['trailer_type'] == 'm3u8_url'){
+                 $data['trailer'] = $data['m3u8_trailer'];
+             }
+             elseif($data['trailer_type'] == 'mp4_url'){
+                 $data['trailer'] = $data['mp4_trailer'];
+             }
+             elseif($data['trailer_type'] == 'embed_url'){
+                 $data['trailer'] = $data['embed_trailer'];
+             }
             
              if(isset($data['duration'])){
                     //$str_time = $data
